+++ vhsdecode/chroma/__init__.py
@@ -5,13 +5,15 @@ importing from vhsdecode.chroma.
 """
 
 from vhsdecode.chroma.qam import (
+    chroma_color_under_filter,
+    decode_chroma,
+    decode_chroma_phase_rotation,
+)
+from vhsdecode.chroma.secam import (
     SECAM_IDENT_MIN_CONFIDENCE,
     SECAM_M1_SEPARATION_RANGE,
     SECAM_M1_UNDER_PAIR_CENTER,
     SecamParityFlywheel,
-    chroma_color_under_filter,
-    decode_chroma,
-    decode_chroma_phase_rotation,
     fit_secam_line_alternation,
     measure_secam_under_carrier_offset,
     regenerate_secam_blanking,
